chore: the student zone sucks btw
This commit marks my great disappointment with the SA team of BukSU. That one personel responded poorly in a very unprofessional manner. The internet is there to serve the users. That's what your Job is, you are literally  paid to do that job. So why cry like a baby? You suck bruh you should do bettter